% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{rrust-package}
\alias{rrust}
\alias{rrust-package}
\title{rrust: Test project for implementing rust code in r packages}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Lynch \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
